requirements.txt
django-environ
pillow
django-crispy-forms
django-allauth
django-filter
gunicorn
whitenoise
django-debug-toolbar
pytest
pytest-django
django-widget-tweaks
django-bootstrap4